Update cstor-pool engine instantiator
This commit will do following:

1. The engine instantiation has been update in accordance
   with latest engine.

2. Remove init status from cstor pool cr.
