+++ cmd/maya-apiserver/spc-watcher/handler.go
@@ -1,9 +1,12 @@
 /*
-Copyright 2018 The OpenEBS Authors.
+Copyright 2017 The OpenEBS Authors
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
+
     http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -53,6 +56,7 @@ func (c *Controller) spcEventHandler(operation string, spcGot *apis.StoragePoolC
 
 		if err !=nil{
 			glog.Error("cstor pool could not be created:",err)
+			// To-Do
 			// If Some error occur patch the spc object with appropriate reason
 		}
